refactor(cms): Moved commands to internal
